The operations of creating the secret and attachedcluster are moved to the suite_test
