+++ e2e/suite_test.go
@@ -18,13 +18,18 @@ package e2e
 
 import (
 	"os"
+	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
 
 	"kurator.dev/kurator/e2e/framework"
+	"kurator.dev/kurator/e2e/resources"
+	clusterv1a1 "kurator.dev/kurator/pkg/apis/cluster/v1alpha1"
 	kurator "kurator.dev/kurator/pkg/client-go/generated/clientset/versioned"
 )
 
@@ -33,6 +38,12 @@ var (
 	kubeClient     kubernetes.Interface
 	kuratorClient  kurator.Interface
 	kuratorContext string
+
+	namespace         string
+	memberClusterName string
+	kubeconfigPath    string
+	secret            *corev1.Secret
+	attachedcluster   *clusterv1a1.AttachedCluster
 )
 
 func TestE2E(t *testing.T) {
@@ -54,4 +65,39 @@ var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
 
 	kuratorClient, err = kurator.NewForConfig(rest)
 	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	namespace = "e2e-test"
+	memberClusterName = "kurator-member"
+	homeDir, err := os.UserHomeDir()
+	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+	kubeconfigPath = filepath.Join(homeDir, ".kube/kurator-member.config")
+
+	// create namespace for e2e test
+	e2eNamespace := resources.NewNamespace(namespace)
+	createNSErr := resources.CreateNamespace(kubeClient, e2eNamespace)
+	gomega.Expect(createNSErr).ShouldNot(gomega.HaveOccurred())
+	time.Sleep(3 * time.Second)
+
+	// build secrets use member cluster kubeconfig
+	kubeconfig, readfileErr := os.ReadFile(kubeconfigPath)
+	gomega.Expect(readfileErr).ShouldNot(gomega.HaveOccurred())
+	data := make(map[string][]byte)
+	data[memberClusterName] = kubeconfig
+	secret = resources.NewSecret(namespace, memberClusterName, data)
+
+	// build attachedclusters for fleet
+	secretKeyRef := clusterv1a1.SecretKeyRef{
+		Name: memberClusterName,
+		Key:  memberClusterName,
+	}
+	attachedcluster = resources.NewAttachedCluster(namespace, memberClusterName, secretKeyRef)
+
+	secretCreateErr := resources.CreateSecret(kubeClient, secret)
+	gomega.Expect(secretCreateErr).ShouldNot(gomega.HaveOccurred())
+
+	attachedCreateErr := resources.CreateAttachedCluster(kuratorClient, attachedcluster)
+	gomega.Expect(attachedCreateErr).ShouldNot(gomega.HaveOccurred())
+	resources.WaitAttachedClusterFitWith(kuratorClient, namespace, memberClusterName, func(attachedCluster *clusterv1a1.AttachedCluster) bool {
+		return attachedCluster.Status.Ready
+	})
 })
